+++ Changelog.md
@@ -1,6 +1,7 @@
 # Changelog
 ## 2019.05.04
 1. Fix Typo at Slogan.vue
+2. Add vuex for keep text while switch layout
 
 ## 2019.03.26